# Delete `EncryptionPerformanceTest`