… Gemfile.lock
… activerecord/test/cases/encryption/performance/encryption_performance_test.rb
… activerecord/test/cases/encryption/performance/envelope_encryption_performance_test.rb
… activerecord/test/cases/encryption/performance/extended_deterministic_queries_performance_test.rb
… activerecord/test/cases/encryption/performance/storage_performance_test.rb