On my machine, running the whole Active Record test suite takes `88` seconds, and `40` of these are spent in encryption tests.

Some of them also happen to occasionally flake because of random blips.

I appreciate the care that has been put into ensuring the overhead of encryption is reasonable, but I don't think these tests justify their cost.

Before:
```
Finished in 87.492219s, 99.8374 runs/s, 339.1844 assertions/s.
8735 runs, 29676 assertions, 0 failures, 0 errors, 33 skips

You have skipped tests. Run with --verbose for details.

Top 10 slowest tests (47.51 seconds, 54.5% of total time):
  ActiveRecord::Encryption::EncryptionPerformanceTest#test_reading_an_encrypted_attribute_multiple_times_is_as_fast_as_reading_a_regular_attribute
    8.3918 seconds test/cases/encryption/performance/encryption_performance_test.rb:22
  ActiveRecord::Encryption::EnvelopeEncryptionPerformanceTest#test_reading_an_encrypted_attribute_multiple_times_is_as_fast_as_reading_a_regular_attribute
    8.3016 seconds test/cases/encryption/performance/envelope_encryption_performance_test.rb:24
  ActiveRecord::Encryption::ExtendedDeterministicQueriesPerformanceTest#test_finding_with_prepared_statement_caching_by_deterministically_encrypted_columns
    6.5172 seconds test/cases/encryption/performance/extended_deterministic_queries_performance_test.rb:7
  ActiveRecord::Encryption::ExtendedDeterministicQueriesPerformanceTest#test_finding_without_prepared_statement_caching_by_encrypted_columns_(deterministic)
    6.4302 seconds test/cases/encryption/performance/extended_deterministic_queries_performance_test.rb:15
  EagerLoadingTooManyIdsTest#test_preloading_too_many_ids
    4.5818 seconds test/cases/associations/eager_test.rb:42
  ActiveRecord::Encryption::EncryptionPerformanceTest#test_performance_when_saving_records
    4.1761 seconds test/cases/encryption/performance/encryption_performance_test.rb:14
  ActiveRecord::Encryption::EnvelopeEncryptionPerformanceTest#test_performance_when_saving_records
    4.1314 seconds test/cases/encryption/performance/envelope_encryption_performance_test.rb:14
  SecurePasswordTest#test_authenticate_by_takes_the_same_amount_of_time_regardless_of_whether_record_is_found
    2.7296 seconds test/cases/secure_password_test.rb:27
  ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_disconnect_and_clear_reloadable_connections_attempt_to_wait_for_threads_to_return_their_conns
    1.2350 seconds test/cases/connection_pool_test.rb:661
  ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_gzip_dumps_identical
    1.0193 seconds test/cases/connection_adapters/schema_cache_test.rb:298

```

After:

```
Finished in 48.428584s, 180.1828 runs/s, 612.1591 assertions/s.
8726 runs, 29646 assertions, 0 failures, 0 errors, 33 skips

You have skipped tests. Run with --verbose for details.

Top 10 slowest tests (11.88 seconds, 24.7% of total time):
  EagerLoadingTooManyIdsTest#test_eager_loading_too_many_ids
    3.8527 seconds test/cases/associations/eager_test.rb:46
  SecurePasswordTest#test_authenticate_by_takes_the_same_amount_of_time_regardless_of_whether_record_is_found
    2.6805 seconds test/cases/secure_password_test.rb:27
  ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_disconnect_and_clear_reloadable_connections_attempt_to_wait_for_threads_to_return_their_conns
    1.2267 seconds test/cases/connection_pool_test.rb:661
  ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_gzip_dumps_identical
    1.0150 seconds test/cases/connection_adapters/schema_cache_test.rb:298
  EagerLoadingTooManyIdsTest#test_preloading_too_many_ids
    0.7487 seconds test/cases/associations/eager_test.rb:42
  ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_checkout_fairness
    0.5638 seconds test/cases/connection_pool_test.rb:416
  ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_checkout_fairness_by_group
    0.5585 seconds test/cases/connection_pool_test.rb:463
  ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_connection_pool_stat
    0.4181 seconds test/cases/connection_pool_test.rb:774
  ActiveRecord::ConnectionAdapters::ConnectionPoolFiberTest#test_connection_pool_stat
    0.4074 seconds test/cases/connection_pool_test.rb:774
  ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_concurrent_connection_establishment
    0.4066 seconds test/cases/connection_pool_test.rb:609
```