@zzak 2024-02-07
@byroot How about just move them to a directory which doesn't run in the same rake task?