@casperisfine 2024-02-07
Feel free to do that if you want. But if you do, try to not have them run on every CI jobs because we're quite short on capacity right now, so I did this to have a better feedback loop in local, but also to reclaim some CI capacity.